Restructure backend definitions
