+++ src/qibo/__init__.py
@@ -1,6 +1,10 @@
 __version__ = "0.1.3-dev"
-from qibo.config import set_precision, set_backend, set_device, set_threads, get_threads, get_backend, get_precision, get_device
-from qibo.config import numpy, matrices, K
+from qibo.config import set_threads, get_threads
+from qibo.backends import set_precision, set_backend, set_device
+from qibo.backends import get_backend, get_precision, get_device
+from qibo.backends import numpy_backend as numpy
+from qibo.backends import numpy_matrices as matrices
+from qibo.backends import K
 from qibo import callbacks
 from qibo import evolution
 from qibo import models
