+++ src/qibo/config.py
@@ -72,87 +72,6 @@ def set_threads(num_threads):
     OMP_NUM_THREADS = num_threads
 
 
-os.environ["TF_CPP_MIN_LOG_LEVEL"] = str(LOG_LEVEL)
-import numpy as np
-import tensorflow as tf
-
-ARRAY_TYPES = (np.ndarray, tf.Tensor)
-NUMERIC_TYPES = (np.int, np.float, np.complex,
-                 np.int32, np.int64, np.float32,
-                 np.float64, np.complex64, np.complex128)
-
-
-from qibo.backends import numpy
-numpy = numpy.NumpyBackend()
-matrices = numpy.matrices
-
-
-from qibo.backends import backend as K
-def set_backend(backend='custom'):
-    """Sets backend used to implement gates.
-
-    Args:
-        backend (str): possible options are 'custom' for the gates that use
-            custom tensorflow operator and 'defaulteinsum' or 'matmuleinsum'
-            for the gates that use tensorflow primitives (``tf.einsum`` or
-            ``tf.matmul`` respectively).
-    """
-    if not ALLOW_SWITCHERS and backend != K.gates:
-        warnings.warn("Backend should not be changed after allocating gates.",
-                      category=RuntimeWarning)
-
-    from qibo.backends import tensorflow
-    bk = tensorflow.TensorflowBackend()
-    K.assign(bk)
-    K.set_gates(backend)
-
-def get_backend():
-    """Get backend used to implement gates.
-
-    Returns:
-        A string with the backend name.
-    """
-    return K.gates
-
-set_backend()
-
-
-def set_precision(dtype='double'):
-    """Set precision for states and gates simulation.
-
-    Args:
-        dtype (str): possible options are 'single' for single precision
-            (complex64) and 'double' for double precision (complex128).
-    """
-    K.precision = dtype
-    matrices.allocate_matrices()
-    tfmatrices.allocate_matrices()
-
-def get_precision():
-    """Get precision for states and gates simulation.
-
-    Returns:
-        A string with the precision name ('single', 'double').
-    """
-    return K.precision
-
-
-def set_device(name):
-    """Set default execution device.
-
-    Args:
-        name (str): Device name. Should follow the pattern
-            '/{device type}:{device number}' where device type is one of
-            CPU or GPU.
-    """
-    K.default_device = name
-    with tf.device(K.default_device):
-        tfmatrices.allocate_matrices()
-
-def get_device():
-    return K.default_device
-
-
 # Configuration for logging mechanism
 t = blessings.Terminal()
