+++ src/qibo/tensorflow/custom_operators/python/ops/qibo_tf_custom_operators.py
@@ -4,7 +4,7 @@ import tensorflow as tf
 import numpy as np
 from tensorflow.python.framework import load_library # pylint: disable=no-name-in-module
 from tensorflow.python.platform import resource_loader # pylint: disable=no-name-in-module
-from qibo.config import get_device, get_threads
+from qibo import get_device, get_threads
 
 
 if tf.config.list_physical_devices("GPU"): # pragma: no cover
